Fix stv0680 camlib compilation warnings.
Resolves a signed-unsigned comparison issue by casting 'for' loop comparison argument to unsigned. This does not change behavior, but clears compiler warning.

+++ camlibs/stv0680/bayer.c
@@ -174,7 +174,7 @@ void light_enhance(int vw, int vh, int coarse, int fine,
 	}
     }
 
-    for (i=0;i<(vw*vh*3);i+=3)
+    for (i=0;i<(unsigned long int)(vw*vh*3);i+=3)
     {
 	int r,g,b;
 	r = *(output+i);
